We write the year 2051. These are the adventure of the Mojito DHT, which, with its 160bit key space is exlploring the limits of the IPv6 address space. In other words I'm using BigIntegers for the estimated DHT size now. As long as the DHT size stays below 2**24 bit we're not using more than 32bit...

+++ components/mojito/src/java/com/limegroup/mojito/MojitoDHT.java
@@ -21,6 +21,7 @@ package com.limegroup.mojito;
 
 import java.io.IOException;
 import java.io.OutputStream;
+import java.math.BigInteger;
 import java.net.InetAddress;
 import java.net.SocketAddress;
 import java.util.Collection;
@@ -130,7 +131,7 @@ public interface MojitoDHT {
     /**
      * Returns the approximate size of the DHT
      */
-    public int size();
+    public BigInteger size();
     
     /**
      * Sets the external (forced) Port number
